backend/requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
pyjwt
azure-messaging-webpubsubservice